[internal/driver/mobile/app] move app#sendLifecycle() and .screenOrientation() into file with more restrictive build flags
These are not used on Windows.
