+++ internal/driver/mobile/app/app.go
@@ -9,7 +9,6 @@ package app
 import (
 	"fyne.io/fyne/v2/internal/async"
 	"fyne.io/fyne/v2/internal/driver/mobile/event/lifecycle"
-	"fyne.io/fyne/v2/internal/driver/mobile/event/size"
 	"fyne.io/fyne/v2/internal/driver/mobile/gl"
 
 	// Initialize necessary mobile functionality, such as logging.
@@ -85,18 +84,6 @@ func init() {
 	theApp.glctx, theApp.worker = gl.NewContext()
 }
 
-func (a *app) sendLifecycle(to lifecycle.Stage) {
-	if a.lifecycleStage == to {
-		return
-	}
-	a.events.In() <- lifecycle.Event{
-		From:        a.lifecycleStage,
-		To:          to,
-		DrawContext: a.glctx,
-	}
-	a.lifecycleStage = to
-}
-
 type app struct {
 	filters []func(any) any
 
@@ -155,12 +142,4 @@ func (a *app) ShowFileOpenPicker(callback func(string, func()), filter *FileFilt
 
 func (a *app) ShowFileSavePicker(callback func(string, func()), filter *FileFilter, filename string) {
 	driverShowFileSavePicker(callback, filter, filename)
-}
-
-func screenOrientation(width, height int) size.Orientation {
-	if width > height {
-		return size.OrientationLandscape
-	}
-
-	return size.OrientationPortrait
 }
